// dawnlight-pages-main/src/types/content.ts
export interface Book {
  slug: string;
  title: string;
  subtitle?: string;
  logline: string;
  cover: string;
  description: string;
  authorSlug: string;
  publishDate: string;
  isbn?: string;
  pages?: number;
  price?: string;
  tableOfContents?: string[];
  purchaseLinks: PurchaseLink[];
}

export interface PurchaseLink {
  name: string;
  url: string;
}

export interface Author {
  slug: string;
  name: string;
  photo?: string;
  bio: string;
  shortBio?: string;
}

export interface NewsArticle {
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  thumbnail?: string;
  publishDate: string;
  category?: string;
}